# БР1.2/Едигарева Дарья/ labs/lab4/services/profiles-service/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Resume": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"profileId": {
						"type": "string"
					},
					"resumePath": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"profileId",
					"resumePath"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResumeCreateDto": {
				"properties": {
					"resumePath": {
						"type": "string"
					}
				},
				"required": [
					"resumePath"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResumeUpdateDto": {
				"properties": {
					"resumePath": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"WorkExperience": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"resumeId": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"resumeId",
					"company",
					"title",
					"startDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExperienceCreateDto": {
				"properties": {
					"company": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"company",
					"title",
					"startDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Education": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"resumeId": {
						"type": "string"
					},
					"institution": {
						"type": "string"
					},
					"degree": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"resumeId",
					"institution",
					"degree",
					"startDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EducationCreateDto": {
				"properties": {
					"institution": {
						"type": "string"
					},
					"degree": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					}
				},
				"required": [
					"institution",
					"degree",
					"startDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JobSeekerProfile": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"userId",
					"fullName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Error": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfileCreateDto": {
				"properties": {
					"fullName": {
						"type": "string"
					}
				},
				"required": [
					"fullName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfileUpdateDto": {
				"properties": {
					"fullName": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "profiles-service",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
		"/me/resumes": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Resume"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Resume"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Resume"
								}
							}
						}
					}
				},
				"tags": [
					"Resume"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResumeCreateDto"
							}
						}
					}
				}
			}
		},
		"/me/resumes/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Resume"
								}
							}
						}
					}
				},
				"tags": [
					"Resume"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResumeUpdateDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"tags": [
					"Resume"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/me/resumes/{id}/experiences": {
			"post": {
				"operationId": "AddExperience",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkExperience"
								}
							}
						}
					}
				},
				"tags": [
					"Resume"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ExperienceCreateDto"
							}
						}
					}
				}
			}
		},
		"/me/resumes/{id}/educations": {
			"post": {
				"operationId": "AddEducation",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Education"
								}
							}
						}
					}
				},
				"tags": [
					"Resume"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EducationCreateDto"
							}
						}
					}
				}
			}
		},
		"/me/profile": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JobSeekerProfile"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JobSeekerProfile"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProfileCreateDto"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JobSeekerProfile"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProfileUpdateDto"
							}
						}
					}
				}
			}
		},
		"/health": {
			"get": {
				"operationId": "Health",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										}
									},
									"required": [
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Health"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "https://[object Object]/"
		}
	]
}